Refactor: Move clickable heading to core components
