+++ src/modules/core/components/ClickableHeading/ClickableHeading.css.d.ts
