+++ src/modules/core/components/ClickableHeading/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ClickableHeading';
