--- src/modules/dashboard/components/ColonyHome/SidebarClickableHeading/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SidebarClickableHeading';
